requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
ca-certificates=2018.1.18=0
cairo=1.14.8=0
certifi=2018.1.18=py27_0
cffi=1.11.2=py27_0
cudatoolkit=8.0=3
cudnn=7.0.5=cuda8.0_0
cycler=0.10.0=py27hc7354d3_0
fontconfig=2.12.1=3
freetype=2.5.5=2
icu=58.2=h9c2bf20_1
intel-openmp=2018.0.0=hc7b2577_8
jpeg=9b=2
libffi=3.2.1=3
libgcc-ng=7.2.0=h7cc24e2_2
libgfortran-ng=7.2.0=h9f7466a_2
libiconv=1.14=0
libpng=1.6.34=0
libstdcxx-ng=7.2.0=hdf63c60_3
libtiff=4.0.9=0
libxml2=2.9.7=h26e45fe_0
matplotlib=1.5.1=np111py27_0
mkl=2018.0.1=h19d6760_4
ncurses=5.9=10
numpy=1.11.3=py27h3dfced4_4
olefile=0.45.1=py27_0
openssl=1.0.2n=0
pillow=4.2.1=py27_0
pip=9.0.1=py27_1
pixman=0.34.0=hceecf20_3
pycairo=1.13.3=py27hea6d626_0
pycparser=2.18=py27_0
pyparsing=2.2.0=py27hf1513f8_1
pyqt=4.10.4=py27_0
pyside=1.2.1=py27_1
python=2.7.14=4
python-dateutil=2.7.0=py27_0
pytorch=0.3.1=py27_cuda8.0.61_cudnn7.0.5_2
pytz=2018.3=py27_0
qt=4.8.5=0
readline=7.0=0
setuptools=38.5.1=py27_0
shiboken=1.2.1=py27_0
sip=4.15.5=py27_0
six=1.11.0=py27_1
sqlite=3.20.1=2
tk=8.6.7=0
torchvision=0.2.0=py27hfb27419_1
wheel=0.30.0=py27_2
xz=5.2.3=0
zlib=1.2.11=0
